requirements.txt
requests>=2.31.0
python-dotenv
pillow
faster-whisper
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
pydantic>=2
pydantic-settings
python-multipart
httpx
huggingface_hub
transformers>=4.40
accelerate
torchvision
bitsandbytes
soundfile
qwen_omni_utils

# Voice Activity Detection
# NOTE: pyaudio requires system package: sudo apt install portaudio19-dev
pyaudio

# Testing & Linting
pytest
pytest-mock
ruff
rich